1. Refactor 2. Project Org 3. Convert to the List
