+++ Apple-Frameworks/Screens/FrameworkDetailView/FrameworkDetailView.swift
@@ -15,19 +15,8 @@ struct FrameworkDetailView: View {
     
     var body: some View {
         VStack {
-            HStack {
-                Spacer()
-                
-                Button {
-                    isShowingDetailView = false
-                } label: {
-                    Image(systemName: "xmark")
-                        .foregroundColor(Color(.label))
-                        .imageScale(.large)
-                        .frame(width: 44, height: 44)
-                }
-            }
-            .padding()
+            
+            XDismissButton(isShowingDetailView: $isShowingDetailView)
             
             Spacer()
