+++ Apple-Frameworks/Views/View/FrameworkTitleView.swift
@@ -1,0 +1,33 @@
+//
+//  FrameworkTitleView.swift
+//  Apple-Frameworks
+//
+//  Created by Юлия Дегтярева on 2025-03-20.
+//
+
+import SwiftUI
+
+struct FrameworkTitleView: View {
+    
+    let framework: Framework
+    
+    var body: some View {
+        VStack {
+            Image(framework.imageName)
+                .resizable()
+                .frame(width: 90, height: 90)
+            Text(framework.name)
+                .font(.title2)
+                .fontWeight(.semibold)
+                .scaledToFit()
+                .minimumScaleFactor(0.6)
+        }
+        .padding()
+    }
+}
+
+struct FrameworkTitleView_Previews: PreviewProvider {
+    static var previews: some View {
+        FrameworkTitleView(framework: MockData.sampleFramework)
+    }
+}
